# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Campus Coin API
  version: 1.0.0-mvp
  license: {name: MIT, identifier: MIT}
  description: |
    Hợp đồng HTTP canonical cho Campus Coin MVP; auth email là register → verify OTP → login → session và forgot password → reset password.
    Google Sign-In là lựa chọn bổ sung qua OIDC; không dùng Gmail credential cá nhân, inbox hoặc Gmail API.
    Mutation cần Origin thuộc allowlist; Referer không thay thế Origin. API response và redirect dùng Cache-Control: no-store, private.
    API không trao quyền tính số dư, authorize payment hoặc ghi tiền cho client/JEV.
servers:
  - url: https://{host}/api/v1
    variables:
      host:
        default: campus-coin.vercel.app
  - url: https://localhost.invalid/api/v1
    description: Môi trường phát triển local; override bằng biến môi trường khi chạy thật
security:
  - sessionCookie: []
...
paths:
  /auth/register:
    post:
      operationId: postAuthRegister
      tags: [Auth]
      security: []
      summary: Gửi OTP để bắt đầu đăng ký bằng email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
          description: Mã OTP đã gửi; tài khoản chưa được tạo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403': {$ref: '#/components/responses/OriginInvalid'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
        '500': {$ref: '#/components/responses/InternalError'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
        default: {$ref: '#/components/responses/InternalError'}
  /auth/verify-registration:
    post:
      operationId: postAuthVerifyRegistration
      tags: [Auth]
      security: []
      summary: Xác minh OTP và tạo tài khoản bằng email, họ tên, mật khẩu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRegistrationRequest'
      responses:
        '200':
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
          description: Xác minh thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403': {$ref: '#/components/responses/OriginInvalid'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
        default: {$ref: '#/components/responses/InternalError'}
  /auth/resend-otp:
    post:
      operationId: postAuthResendOtp
      tags: [Auth]
      security: []
      summary: Gửi lại mã OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendOtpRequest'
      responses:
        '200':
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
          description: OTP đã gửi lại
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403': {$ref: '#/components/responses/OriginInvalid'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
        '500': {$ref: '#/components/responses/InternalError'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
        default: {$ref: '#/components/responses/InternalError'}
  /auth/login:
    post:
      operationId: postAuthLogin
      tags: [Auth]
      security: []
      summary: Đăng nhập email và mật khẩu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
          description: Đăng nhập thành công, thiết lập session cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/LoginForbidden'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
        default: {$ref: '#/components/responses/InternalError'}
  /auth/providers:
    get:
      operationId: getAuthProviders
      tags: [Auth]
      security: []
      summary: Lấy các phương thức đăng nhập đang được cấu hình
      responses:
        '200': {description: Phương thức hiện có, headers: {Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}}, content: {application/json: {schema: {$ref: '#/components/schemas/AuthProvidersResponse'}}}}
        '500': {$ref: '#/components/responses/InternalError'}
        default: {$ref: '#/components/responses/InternalError'}
  /auth/google/start:
    get:
      operationId: getAuthGoogleStart
      tags: [Auth]
      security: []
      summary: Bắt đầu Google Sign-In bằng OAuth Authorization Code và PKCE
      responses:
        '302': {description: Chuyển trình duyệt đến Google; cookie flow ngắn hạn được thiết lập, headers: {Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}, Location: {schema: {type: string, format: uri}}, Set-Cookie: {schema: {type: string}}}}
        default: {description: Lỗi chuyển hướng đã được xử lý an toàn, headers: {Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}}}
  /auth/google/link:
    post:
      operationId: postAuthGoogleLink
      tags: [Auth]
      summary: Bắt đầu kết nối Google từ tài khoản đã đăng nhập
      x-csrf-required: true
      responses:
        '200': {description: URL xác thực Google và cookie flow, headers: {Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}, Set-Cookie: {schema: {type: string}}}, content: {application/json: {schema: {$ref: '#/components/schemas/GoogleLinkStartResponse'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/CsrfInvalid'}
        '409': {$ref: '#/components/responses/Conflict'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
        '500': {$ref: '#/components/responses/InternalError'}
        default: {$ref: '#/components/responses/InternalError'}
  /auth/google/callback:
    get:
      operationId: getAuthGoogleCallback
      tags: [Auth]
      security: []
      summary: Hoàn tất callback OIDC Google và chuyển về ứng dụng
      parameters:
        - {name: state, in: query, required: true, schema: {type: string}}
        - {name: code, in: query, schema: {type: string}}
        - {name: error, in: query, schema: {type: string}}
      responses:
        '302': {description: Chuyển về ứng dụng; thiết lập session khi đăng nhập thành công, headers: {Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}, Location: {schema: {type: string, format: uri}}, Set-Cookie: {schema: {type: string}}}}
        default: {description: Lỗi callback được chuyển về ứng dụng bằng mã lỗi an toàn, headers: {Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}}}
  /auth/forgot-password:
    post:
      operationId: postAuthForgotPassword
      tags: [Auth]
      security: []
      summary: Yêu cầu OTP đặt lại mật khẩu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
          description: Phản hồi tiêu chuẩn không để lộ sự tồn tại của email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403': {$ref: '#/components/responses/OriginInvalid'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '500': {$ref: '#/components/responses/InternalError'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
        default: {$ref: '#/components/responses/InternalError'}
  /auth/reset-password:
    post:
      operationId: postAuthResetPassword
      tags: [Auth]
      security: []
      summary: Đặt lại mật khẩu với OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
          description: Đặt lại mật khẩu thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403': {$ref: '#/components/responses/OriginInvalid'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
        default: {$ref: '#/components/responses/InternalError'}
  /auth/csrf:
    get:
      operationId: getAuthCsrf
      tags: [Auth]
      summary: Lấy CSRF token
      responses:
        '200':
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
          description: Token CSRF
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsrfTokenResponse'
        '401': {$ref: '#/components/responses/Unauthorized'}
        '500': {$ref: '#/components/responses/InternalError'}
        default: {$ref: '#/components/responses/InternalError'}
  /users/me:
    get:
      operationId: getUsersMe
      tags: [Auth]
      summary: Lấy thông tin user hiện tại (owner scope)
      responses:
        '200':
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
          description: Thông tin user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401': {$ref: '#/components/responses/Unauthorized'}
        '500': {$ref: '#/components/responses/InternalError'}
        default: {$ref: '#/components/responses/InternalError'}
...
  /auth/session:
    get:
      operationId: getAuthSession
      tags: [Auth]
      summary: Lấy người dùng và CSRF token của phiên hiện tại
      responses:
        '200':
          description: Session hợp lệ
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
          content: {application/json: {schema: {$ref: '#/components/schemas/SessionResponse'}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '500': {$ref: '#/components/responses/InternalError'}
        default: {$ref: '#/components/responses/InternalError'}
  /auth/logout:
    post:
      operationId: postAuthLogout
      tags: [Auth]
      summary: Revoke session hiện tại; yêu cầu CSRF khi phiên còn hiệu lực, thao tác lặp là idempotent
      x-csrf-required: true
      responses:
        '200': {description: Session đã revoke và cookie được clear, headers: {Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}}, content: {application/json: {schema: {$ref: '#/components/schemas/MessageResponse'}}}}
        '403': {$ref: '#/components/responses/CsrfInvalid'}
        '500': {$ref: '#/components/responses/InternalError'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
        default: {$ref: '#/components/responses/InternalError'}
  /users/me/preferences:
    patch:
      operationId: patchUsersMePreferences
      tags: [Auth]
      summary: Cập nhật preference người dùng
      x-csrf-required: true
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UpdatePreferencesRequest'}}}}
      responses:
        '200': {description: Preference đã cập nhật, headers: {Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}}, content: {application/json: {schema: {$ref: '#/components/schemas/UserResponse'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/CsrfInvalid'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '500': {$ref: '#/components/responses/InternalError'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
        default: {$ref: '#/components/responses/InternalError'}
...
  /issues:
    post:
      operationId: postIssues
      tags: [Issues]
      summary: Tạo issue hỗ trợ
      x-csrf-required: true
      x-idempotency-required: true
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateIssueRequest'}}}}
      responses:
        '201': {description: Issue đã tạo, headers: {Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}}, content: {application/json: {schema: {$ref: '#/components/schemas/IssueResponse'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/CsrfInvalid'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/IdempotencyConflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '500': {$ref: '#/components/responses/InternalError'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
        default: {$ref: '#/components/responses/InternalError'}
  /issues/me:
    get:
      operationId: getIssuesMe
      tags: [Issues]
      summary: Liệt kê issue của user hiện tại
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses:
        '200': {description: Issue page, headers: {Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}}, content: {application/json: {schema: {$ref: '#/components/schemas/IssuePageResponse'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '500': {$ref: '#/components/responses/InternalError'}
        default: {$ref: '#/components/responses/InternalError'}
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [Admin]
      summary: Thống kê số lượng user đã xác minh
      responses:
        '200': {description: User statistics, headers: {Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}}, content: {application/json: {schema: {$ref: '#/components/schemas/AdminStatsResponse'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '500': {$ref: '#/components/responses/InternalError'}
        default: {$ref: '#/components/responses/InternalError'}
...
components:
  headers:
    PrivateNoStore:
      description: Không lưu response riêng tư trong browser hoặc shared cache.
      schema:
        type: string
        example: no-store, private
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cc_session
      description: Opaque server-side session cookie; không phải JWT.
...
  responses:
...
    OriginInvalid: {description: Origin của request không được phép, headers: {Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}}, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
  schemas:
...
    RegisterRequest:
      type: object
      required: [email]
      properties:
        email: {type: string, format: email}
...
    MessageResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [message]
          properties:
            message: {type: string}
...